==> laboratory/legacy/samples/01_idea.ipynb <==
# Dig into Dataframe

import pandas as pd
import numpy as np
import pandas_datareader as pdr
import investpy as iv
import pandas_ta as ta
from datetime import datetime as dt
pd.set_option('max_columns', 6, 'max_rows', 6, 'display.precision', 4)

# quotes = 'GBPUSD=X'
# data_yh = pdr.DataReader(quotes, data_source="yahoo",
#                       start='01/01/2020', end='12/31/2021')
# data_yh.to_csv('..\\cookbook_data\\owndata\\GBPUSD_yh.csv')

# data_yh = pd.read_csv('..\\cookbook_data\\owndata\\GBPUSD_yh.csv',
#                       index_col=0, parse_dates=True)
# data_yh.head()


# data_iv = iv.currency_crosses.get_currency_cross_historical_data('GBP/USD', '01/01/2019', '03/02/2022')
# data_iv.to_csv('..\\cookbook_data\\owndata\\GBPUSD_iv.csv')

# data_iv = pd.read_csv('..\\data\\owndata\\GBPUSD_iv.csv',
#                       index_col=0, parse_dates=True)
# data_iv.tail()


quotes = ('PHLX Australian Dollar', 'PHLX Canadian Dollar',
          'PHLX Euro', 'PHLX Yen', 'PHLX New Zealand Dollar',
          'PHLX Swiss Franc', 'PHLX British Pound',
          'US Dollar Index')

names = [quote.replace(' ', '_') for quote in quotes]
# names

today = dt.today().strftime('%d/%m/%Y')
# today

quotes = ('US Dollar Index',)

for quote in quotes:
    data_iv = iv.indices.get_index_historical_data(quote, 'united states', '01/01/2019', today)
    name = quote.replace(' ', '_')
    data_iv.to_csv(f'..\\cookbook_data\\owndata\\{name}_iv.csv')


# End Extract Data

name = names[3]
name

data_iv = pd.read_csv(f'..\\cookbook_data\\owndata\\{name}_iv.csv',
                      index_col=0, parse_dates=True)
data_iv.tail()


data_iv['ohlc4'] = ta.ohlc4(data_iv.Open, data_iv.High, data_iv.Low, data_iv.Close)
# data_iv.drop(['Currency'], axis=1, inplace=True)      # because notebook save file status
data_iv.head()

columns_iv = data_iv.columns
index_iv = data_iv.index
values_iv = data_iv.values

# columns_iv
print(type(columns_iv))

# index_iv
print(type(index_iv))

# values_iv
print(type(values_iv))

columns_iv.values

# index_iv.values

#### Data types

data_iv = pd.read_csv(f'..\\data\\owndata\\{name}_iv.csv',
                      index_col=0, parse_dates=True)
data_iv['hl2'] = ta.hl2(data_iv.High, data_iv.Low)
data_iv['hlc3'] = ta.hlc3(data_iv.High, data_iv.Low, data_iv.Close)
data_iv['ohlc4'] = ta.ohlc4(data_iv.Open, data_iv.High, data_iv.Low, data_iv.Close)
# data_iv.drop(['Currency'], axis=1, inplace=True)      # because notebook save file status
# data_iv.drop(['Currency', 'Volume'], axis=1, inplace=True)      # because notebook save file status
data_iv.dtypes

### [New idea: Fear & Greed Index](https://money.cnn.com/data/fear-and-greed)
- [Safe haven assets](https://www.venn.twosigma.com/vennsights/five-safe-haven-assets)
- How to calculate ... 

data_iv.dtypes.value_counts()

data_iv['ohlc4_change'] = data_iv.ohlc4.pct_change().abs()

data_iv.tail()

# Selecting 

ohlc4 = data_iv.ohlc4
ohlc4.name

ohlc4.to_frame().tail()

new_data_iv = data_iv.iloc[1:]
# new_data_iv[-52:]
new_data_iv[-100:].describe()

ohlc4_change = new_data_iv.ohlc4_change

ohlc4_change.shape, ohlc4_change.size, len(ohlc4_change), ohlc4_change.count()

pd.set_option('display.precision', 6)

# ohlc4_change.describe() 

norm_ohlc4_change = ohlc4_change.describe() * 100
norm_ohlc4_change

norm_ohlc4_change.quantile(.2)

norm_ohlc4_change.quantile([.1, .2, .3, .4, .5, .6, .7, .8])

ohlc4_change.isnull().value_counts()

### add daily min risk depend on min / mean value of ohlc change

norm_ohlc4_change

temp_data = ohlc4_change[-50:]
print(temp_data.describe())

mean_ohlc4_change = temp_data.mean()
mean_ohlc4_change * 100

count_above_mean = temp_data >= mean_ohlc4_change
count_above_mean.value_counts()

data_iv = pd.read_csv(f'..\\cookbook_data\\owndata\\{name}_iv.csv',
                      index_col=0, parse_dates=True)
data_iv['hl2'] = ta.hl2(data_iv.High, data_iv.Low)
data_iv['hlc3'] = ta.hlc3(data_iv.High, data_iv.Low, data_iv.Close)
data_iv['ohlc4'] = ta.ohlc4(data_iv.Open, data_iv.High, data_iv.Low, data_iv.Close)
# data_iv.drop(['Currency'], axis=1, inplace=True)      # because notebook save file status
data_iv

data_iv['hl2_change'] = data_iv.hl2.pct_change().abs()
data_iv['hlc3_change'] = data_iv.hlc3.pct_change().abs()
data_iv['ohlc4_change'] = data_iv.ohlc4.pct_change().abs()
data_iv

temp_data = data_iv[-50:]
temp_data_index = temp_data.index
# temp_data.describe()

mean_hl2_change = temp_data.hl2_change.mean()
mean_hlc3_change = temp_data.hlc3_change.mean()
mean_ohlc4_change = temp_data.ohlc4_change.mean()

hl2_above_mean = temp_data.hl2_change >= mean_hl2_change
hl2_above_mean.value_counts()
# hl2_above_mean
# type(hl2_above_mean)

np_hl2_above_mean = hl2_above_mean.to_numpy()

hlc3_above_mean = temp_data.hlc3_change >= mean_hlc3_change
hlc3_above_mean.value_counts()
# hlc3_above_mean

np_hlc3_above_mean = hlc3_above_mean.to_numpy()

ohlc4_above_mean = temp_data.ohlc4_change >= mean_ohlc4_change
ohlc4_above_mean.value_counts()
# ohlc4_above_mean

np_ohlc4_above_mean = ohlc4_above_mean.to_numpy()

import numpy as np

np_result = np.logical_and(np.logical_and(np_hl2_above_mean, np_hlc3_above_mean), np_ohlc4_above_mean)

pd_result = pd.Series(np_result, index=temp_data_index)
pd_result.value_counts()

data_iv['close_change'] = data_iv.Close.pct_change().abs()
data_iv['close_status'] = data_iv.Close.pct_change()

sample = data_iv[-50:]
sample_index = sample.index

mean_close_change = sample.close_change.mean()

col_len = len(sample.columns)
col_len

# close_above_mean = sample.close_change >= mean_close_change

# close_above_mean = sample.close_change.gt(mean_close_change)
# close_above_mean.value_counts()

sample['gt_mean_abs'] = sample.close_change.gt(mean_close_change)
sample

# import warnings
sample['status'] = sample.close_status.gt(0.0)
# sample[['gt_mean_abs', 'status']]

# (greater/ smaller) and (bull or bear)
sample['final_status'] = sample.gt_mean_abs.astype(int) * sample.status.astype(int)

result_stt_sr = sample['final_status'].astype(bool)
result_stt_sr.value_counts(normalize=True)

sample.gt_mean_abs.value_counts()   # mean 28 times change < mean, 22 times change > mean 

sample.status.value_counts()    # mean 28 times bull, 22 bear

## Volume adding: T.B.D
- yahoo finance refer
- change > 75% ...
