==> laboratory/legacy/samples/01_shit.ipynb <==
import pandas as pd
import numpy as np
import pandas_datareader as pdr
import investpy as iv
import pandas_ta as ta
pd.set_option('max_columns', 8, 'max_rows', 10, 'display.precision', 4)

quotes = ('PHLX Australian Dollar', 'PHLX Canadian Dollar',
          'PHLX Euro', 'PHLX Yen', 'PHLX New Zealand Dollar',
          'PHLX Swiss Franc', 'PHLX British Pound',
          'US Dollar Index')

names = [quote.replace(' ', '_') for quote in quotes]
# names

def get_dataiv(filepath, index=0, parsed=True):
    try:
        data = pd.read_csv(filepath, index_col=index, parse_dates=parsed)
        data.drop(['Volume'], axis=1, inplace=True) if 'Volume' in data_iv.columns else None
        data.drop(['Currency'], axis=1, inplace=True) if 'Currency' in data_iv.columns else None
    except FileNotFoundError:
        print("File not found. Please check the filepath.")
    return data


# 'Volume' in data_iv.columns

# 'Currency' in data_iv.columns

df = get_dataiv(f'..\\cookbook_data\\owndata\\{names[-1]}_iv.csv')
df

# data_iv = pd.read_csv(f'..\\cookbook_data\\owndata\\{names[-1]}_iv.csv')
# data_iv.tail()

df.shape, df.columns, df.index

# df.info()
# df.describe()

df.dtypes

df.dtypes.value_counts()

close = df['Close']
type(close), close.name

to_dframe = close.to_frame()
type(to_dframe)

to_dframe.tail()

#### some common functions, prop:
- head, tail, drop, to_frame, value_counts, count, describe, quantile, fillna, 
- size, shape, dtypes, columns, index

close.hasnans

df['wma'] = ta.wma(close, timeperiod=20)
df.dropna(inplace=True)
df

df.Close.add(data_iv.wma)

df['Square'] = df.Close.mul(df.wma).pow(0.25)
df['M_Square'] = df.Close.pow(0.5).ge(df.Square)
df
