==> laboratory/legacy/samples/02_main_sentiment.ipynb <==
import pandas as pd
import numpy as np
pd.set_option('max_columns', 15, 'max_rows', 30, 'display.precision', 4)

prices_name = 'price2302_df'
sentiments_name = 'sent2302_df'

prices = pd.read_csv(f'../cookbook_data/owndata/{prices_name}.csv', index_col=0)
sentiments = pd.read_csv(f'../cookbook_data/owndata/{sentiments_name}.csv', index_col=0)

prices.rename(columns={'Avg. Short Price /  Distance From Price': 'Short_Distance',
                       'Avg. Long Price /  Distance From Price': 'Long_Distance',
                       'Current Price': 'Current_Price'}, inplace=True)

distance = prices.iloc[:, 0:2].copy()

distance.Short_Distance = distance.Short_Distance.map(lambda x: x.rstrip(' pips'))
distance.Long_Distance = distance.Long_Distance.map(lambda x: x.rstrip(' pips'))

temp_short = distance.Short_Distance.str.split(
    ' ', expand=True).iloc[:, [0, 2]].copy()
temp_short.rename(columns={
                  temp_short.columns[0]: 'Short_Price', temp_short.columns[1]: 'Short_Distance'}, inplace=True)
# temp_short


temp_long = distance.Long_Distance.str.split(
    ' ', expand=True).iloc[:, [0, 2]].copy()
temp_long.rename(columns={
    temp_long.columns[0]: 'Long_Price', temp_long.columns[1]: 'Long_Distance'}, inplace=True)
# temp_long


result = pd.concat([temp_short, temp_long, prices.Current_Price], axis=1)
# result
# result.to_csv('../cookbook_data/owndata/prices_cleandf.csv')

# result.dtypes.value_counts()

result = result.apply(pd.to_numeric)
# result.dtypes

# result.select_dtypes(include=['int'])

# result.select_dtypes(include=['float'])

# result.select_dtypes(include=['number'])

##### addition functions

def max_streak(s):
    s1 = s.cumsum()
    return s.mul(s1).diff().where(lambda x: x < 0).ffill().add(s1, fill_value=0).max()

# weekly weight loss updates
def find_perc_loss(s):
    ''' calculate percent loss in weight '''
    return (s - s.iloc[0]) / s.iloc[0]


# generic closure to build all of our customized functions
def make_agg_func(func, name, *args, **kwargs):
    def wrapper(x):
        # print(type(x))    # x ~ s: Series performs function
        return func(x, *args, **kwargs)
    wrapper.__name__ = name
    return wrapper

# my_agg1 = make_agg_func(pct_between, 'pct_1_3k', low=1000, high=3000)
# my_agg2 = make_agg_func(pct_between, 'pct_10_30k', 10000, 30000)

def pct_between(s, low, high):
    return s.between(low, high).mean()

def max_deviation(s):
    std_score = (s - s.mean()) / s.std()
    return std_score.abs().max()

def test_return_normality(stock_data):
    close = stock_data['Close']
    daily_return = close.pct_change().dropna()
    daily_return.hist(bins=20)
    mean = daily_return.mean()
    std = daily_return.std()

    abs_z_score = abs(daily_return - mean) / std

    pcts = [abs_z_score.lt(i).mean() for i in range(1, 4)]
    for count, pct in enumerate(pcts, 1):
        print(f'{pct:.3f} fall within {count}')

##### end add function

# result.index

result['Relative_SL'] = result.Short_Distance / result.Long_Distance
result['RelativePrice_SL'] = result.Short_Price / result.Long_Price

result.sort_values(by='Short_Distance', ascending=False, key=abs).head()

result.sort_values(by='Long_Distance', ascending=False, key=abs).head()

result.sort_values(by='Relative_SL', ascending=False, key=abs).head()

result.sort_values(by='Relative_SL', ascending=True, key=abs).head()

# balance or not balance
result.sort_values(by='RelativePrice_SL', ascending=False).head()

result.sort_values(by='RelativePrice_SL', ascending=True).head()

# sorted_result.head()

# sorted_result.tail()

# result.select_dtypes(include=['string'])

# need update current price
# result.filter(items=['Short_Price', 'Long_Price', 'Current_Price']).head()
result.filter(items=['Short_Price', 'Long_Price']).head()

result.filter(like='Distance').head()

# result.filter(regex='\d').head()
# result.columns

# result.values

# result.count()

# result.min()

# result.max()

# result.describe()

# pd.set_option('max_rows', 20)
# result.describe(percentiles=[.01, .05, .1, .3, .4, .5, .6, .7, .8, .9, .95, .99])

# pd.set_option('max_rows', 10)
# result.isnull().sum()

# result.min(skipna=False)

# result.isnull().sum().sum()

# result.isnull().any()

# result.isnull().dtypes

# result.dtypes
# result.select_dtypes(['int']).fillna('').max()
result.select_dtypes(['int']).fillna('').sort_values(by=['Short_Distance'], ascending=False).tail(10)   # head

result.Long_Distance.sort_values(ascending=False).head(10)              

#### update day by day this price

# result

# sentiments = sentiments.apply(pd.to_numeric)

sentiments.dtypes

sentiments.head(10)

sentiments.tail(10)

## First Example

status = sentiments.iloc[:, 0].copy()
# status

details_copy = sentiments.iloc[:, 1:4].copy()
details_copy.Volume = details_copy.Volume.map(lambda x: x.rstrip(' lots'))
details_copy.Percentage = details_copy.Percentage.map(lambda x: x.rstrip('%'))
details_copy = details_copy.apply(pd.to_numeric)

details = pd.concat([status, details_copy], axis=1)

details.head(10)

# details.sort_values(by=['Volume'], ascending=False).head(10)   # head

details.sort_values(by=['Percentage'], ascending=False).head(10)

# details.sort_values(by=['Percentage'], ascending=False).tail(10)

quote = r'GBP'
quote_filter = details[details.index.str.contains(quote + r'(?!$)')]
quote_filter.dtypes
# quote_filter = quote_filter.apply(pd.to_numeric)

quote_filter.sort_values(by=['Percentage'], ascending=False)

quote = r'CHF'
quote_filter = details[details.index.str.contains(quote + r'(?!$)')]
quote_filter.dtypes
# quote_filter = quote_filter.apply(pd.to_numeric)

quote_filter.sort_values(by=['Percentage'], ascending=False)

## More from sentiments

sentiments.Volume = sentiments.Volume.map(lambda x: x.rstrip(' lots'))
sentiments.Percentage = sentiments.Percentage.map(lambda x: x.rstrip('%'))

sentiments

numeric_sentiments = sentiments.iloc[:, 1:4].apply(pd.to_numeric)

result_sentiments = pd.concat([sentiments.Action, numeric_sentiments], axis=1)

result_sentiments
# result_sentiments.to_csv('../cookbook_data/owndata/sentiments_df.csv')

result_sentiments.sort_values(by=['Volume'], ascending=False).head(10)

result_sentiments.sort_values(by=['Positions'], ascending=False).head(10)

#### Sum Positions by Symbol

#### Diff Percentage|Positions|Volume by Symbol

#### Merge Symbols and Actions

#### Compare Volume betwwen Quotes (with same Base): ... 

economic_calendar_1 = pd.read_csv('../cookbook_data/owndata/economic_calendar_1.csv', index_col=1)
economic_calendar_1.drop('id', axis=1, inplace=True)
time_format = '%d/%m/%Y'
economic_calendar_1.index = pd.to_datetime(economic_calendar_1.index, format=time_format)

notnan_economic_filter = economic_calendar_1[economic_calendar_1.currency.notnull()]
# notnan_economic_filter

pd.set_option('max_columns', 20, 'max_rows', 20)

quote_filter = notnan_economic_filter[notnan_economic_filter.currency.str.contains('JPY')]
quote_filter = quote_filter.dropna()

def strip_shit(df, cols=['actual', 'forecast', 'previous']):
    for col in cols:
        df[col] = df[col].map(lambda x: x.rstrip(
            '%').rstrip('T').rstrip('B').rstrip('K').rstrip('M'))
    return df

quote_filter.iloc[:, 5:8] = strip_shit(quote_filter.iloc[:, 5:8].copy())
quote_result_num = quote_filter.iloc[:, 5:8].apply(pd.to_numeric)
quote_results = pd.concat([quote_filter.iloc[:, :5], quote_result_num], axis=1)
quote_results.dtypes

quote_results.select_dtypes(include=['float']).head()

quote_results.select_dtypes(include=['object']).head()

#### NLP some event words: YoY, MoM
- Calculate months range, monthly change, yearly changes
- Remove shitty words such as Q1

#### [About Fx Market Holiday](https://www.forex.academy/how-do-public-holidays-affect-the-forex-market/)

economic_calendar_1

# first way to filter holiday 
holiday_filter = economic_calendar_1[economic_calendar_1.currency.isnull()]

# holiday_filter.to_csv('../cookbook_data/owndata/economic_calendar_1_holiday.csv')
holiday_filter

len(economic_calendar_1), len(holiday_filter), len(notnan_economic_filter)

#### events detail: (Split to Nation)
- Housing: HPI / Building Approvals
- Retail Sales
- CPI / HIPC / Core CPI
- PMI: Services / Markit Composite / Manufacturing / Composite 
- Unemployment Change / Unemployment Rate
- Speaks: RBA
- Business Confidence
- Trade Balance
- Inflation Report
- MPC: Vote Cut

#### GBP Examples
- Housing: [HPI](https://www.investing.com/economic-calendar/nationwide-hpi-850), [Halifax](https://www.investing.com/economic-calendar/halifax-house-price-index-844) House Price Index, [RICS](https://www.investing.com/economic-calendar/rics-house-price-balance-264) House Price Balance
- PMI: [Manufacturing](https://www.investing.com/economic-calendar/manufacturing-pmi-204) ... [Composite](https://www.investing.com/economic-calendar/composite-pmi-1934), [Services](https://www.investing.com/economic-calendar/services-pmi-274), [Construction](https://www.investing.com/economic-calendar/construction-pmi-44)
- [Inflation](https://www.investing.com/economic-calendar/boe-inflation-report-15) 
- [PPI](https://www.investing.com/economic-calendar/ppi-input-242)
- MPC, BOE Member Speech
- [Interest Rate](https://www.investing.com/economic-calendar/interest-rate-decision-170) Decision
- Governor Speaks
- [CFTC](https://www.investing.com/economic-calendar/cftc-gbp-speculative-positions-1612) GBP speculative net positions
- GBP/USD [Index](https://www.investing.com/economic-calendar/investing.com-gbp-usd-index-1156)
- [BRC](https://www.investing.com/economic-calendar/brc-retail-sales-monitor-18) [Retail Sales](https://www.investing.com/economic-calendar/retail-sales-731), [Core](https://www.investing.com/economic-calendar/core-retail-sales-857)
- [Business Investment](https://www.investing.com/economic-calendar/business-investment-30)
- [GDP](https://www.investing.com/economic-calendar/gdp-121), [Monthly](https://www.investing.com/economic-calendar/monthly-gdp-3m-3m-change-1800)
- Production [Industrial](https://www.investing.com/economic-calendar/industrial-production-158), [Manufacturing](https://www.investing.com/economic-calendar/manufacturing-production-205)
- [Trade Balance](https://www.investing.com/economic-calendar/trade-balance-284), [Non-EU](https://www.investing.com/economic-calendar/trade-balance-non-eu-849)
- [NIESR](https://www.investing.com/economic-calendar/niesr-monthly-gdp-tracker-1797)
- [Average Earnings Index +Bonus](https://www.investing.com/economic-calendar/average-earnings-index-bonus-7)
- [Claimant Count Change](https://www.investing.com/economic-calendar/claimant-count-change-39)
- [Employment](https://www.investing.com/economic-calendar/employment-change-3m-3m-1743), [Unemployment](https://www.investing.com/economic-calendar/unemployment-rate-297)
- [CBI](https://www.investing.com/economic-calendar/cbi-industrial-trends-orders-34) Industrial Trends Orders

economic_calendar_2 = pd.read_csv('../cookbook_data/owndata/economic_calendar_2.csv', index_col=1)
economic_calendar_2.drop('id', axis=1, inplace=True)

time_format = '%d/%m/%Y'
economic_calendar_2.index = pd.to_datetime(economic_calendar_2.index, format=time_format)

def quote_filtering(df, quote):
    pd.set_option('max_columns', 20, 'max_rows', 20)

    quote_filter = df[df.currency.str.contains(quote)]
    quote_filter_medium = quote_filter[quote_filter.importance.str.contains('medium')]
    quote_filter_high = quote_filter[quote_filter.importance.str.contains('high')]
    return quote_filter_medium, quote_filter_high

def quote_numericing(df):
    df.iloc[:, 5:8] = strip_shit(df.iloc[:, 5:8].copy())
    quote_result_num = df.iloc[:, 5:8].apply(pd.to_numeric)
    return pd.concat([df.iloc[:, :5], quote_result_num], axis=1)


def split_filter(df):
    return df[df.isna().any(axis=1)], quote_numericing(df.copy().dropna())


notnan_economic_filter_2 = economic_calendar_2[economic_calendar_2.currency.notnull()]

quote_filter_medium, quote_filter_high = quote_filtering(notnan_economic_filter_2, 'USD')
quote_notall_num, result_numeric = split_filter(quote_filter_high)

result_numeric

#### dropna make some case not removing cell number Unit 

quote_notall_num
