==> laboratory/legacy/samples/03_shit.ipynb <==
# Beginning Data Analysis with real financial data

import numpy as np
import pandas as pd
from IPython.display import display
pd.set_option('display.max_columns', 20)    # , 'display.max_rows', 20

dxy = pd.read_csv('../cookbook_data/owndata/US_Dollar_Index_iv.csv', index_col=0, parse_dates=True)
# type(dxy.index)

dxy.drop(['Volume', 'Currency'], axis=1, inplace=True)

# dxy.head()

dxy.shape

num = 30
tmp_dxy = dxy.tail(num)
# tmp_dxy

with pd.option_context('display.max_rows', 8):
    # display(dxy.describe(include=[np.number]).T)

    # refer
    # result.describe(percentiles=[.01, .05, .1, .3, .4, .5, .6, .7, .8, .9, .95, .99])

    display(tmp_dxy.describe(include=[np.number],
            percentiles=[.05, .25, .5, .75, .95]).T)


tmp_dxy.dtypes

# tmp_dxy.describe(include=[object, pd.Categorical]).T
