==> laboratory/legacy/samples/04_calendar.ipynb <==
# Selecting Subsets of Real financial Data

import pandas as pd
import numpy as np
from data_pre import (
    strip_unuse, quote_filtering, quote_numericing, split_filter)
pd.set_option('display.max_rows', 8)


# import sys
# sys.path.insert(0, '..')

#### read data path

data = pd.read_csv('../cookbook_data/owndata/economic_calendar_2.csv', index_col=1)
data.drop('id', axis=1, inplace=True)

time_format = '%d/%m/%Y'
data.index = pd.to_datetime(data.index, format=time_format)

data.head()

events = data['event']
events.head(11)

events.iloc[2]

events.iloc[[1, 5, 10]]

events.iloc[1:10:2]

notnan_filter = data[data.currency.notnull()]
quote_medium, quote_high = quote_filtering(notnan_filter, 'USD')
notall_num, result_num = split_filter(quote_medium)
# result_num

# events.loc['2022-02-11']
result_num.loc['2022-02-01']

np.random.seed(2)

labels = list(np.random.choice(events.index, 1))      # random sampling
labels

events.loc[labels]

result_num['2022-02-01': '2022-02-03']

result_num['2022-02-01': '2022-02-03': 2]

result_num.iloc[[2]]

result_num.loc['2022-02-01': '2022-02-03': 2]

result_num['2022-02-04': '2022-02-01': -2]

result_num.iloc[:3, 4:]    # first 3 rows, last 4 columns

result_num.iloc[:, [4, 5, 6, 7]].head()
# result_num.iloc[:, 4:8].head()    # same above

result_num.iloc[[2, 8], [6, 7]]

result_num.iloc[5, -4] # get cell at row 5, column -4
# result_num.iloc[:, -4]

result_num.iloc[10:3:-2, 4]   # get cells at rows 90 to 80, column 5

result_num.loc[:'2022-02-03', :'time']

result_num.loc[:, ['time', 'event']]

result_num.loc['2022-02-01', 'event']

type(result_num.index)

# rows = ['2022-02-01', '2022-02-02', '2022-02-03']
# time_rows = list(pd.DatetimeIndex([pd.to_datetime(row, format='%Y-%m-%d')]) for row in rows)
# print(time_rows)

# columns = ['forecast', 'previous']
# result_num.loc[time_rows, columns]

result_event = result_num.copy().set_index('event')
# result_event.loc['Manufacturing PMI  (Jan)']
# result_event.index
# type(result_event.index)

# ------------ pass name instead of index ------------
rows = ['Manufacturing PMI  (Jan)', 'Average Hourly Earnings (MoM)  (Jan)', 'API Weekly Crude Oil Stock']
columns = ['forecast', 'previous']
result_event.loc[rows, columns]

result_event.loc['Manufacturing PMI  (Jan)', 'forecast']

result_event.index

start, stop = 'Manufacturing PMI  (Jan)', 'Private Nonfarm Payrolls  (Jan)'
result_event.loc[start:stop:2, 'forecast']


## Selecting data with both integers and labels

data = pd.read_csv('../cookbook_data/owndata/economic_calendar_2.csv', index_col=1)
data.drop('id', axis=1, inplace=True)

time_format = '%d/%m/%Y'
data.index = pd.to_datetime(data.index, format=time_format)

col_start = data.columns.get_loc('event')
col_end = data.columns.get_loc('event') + 4

data.iloc[:5, col_start:col_end]

len(data.index)

#### unique index issue

# data.index = data.index.unique()
# data

data_event = data.copy().set_index('event')

#### [count the duplicates of each type of row](https://stackoverflow.com/questions/35584085/how-to-count-duplicate-rows-in-pandas-dataframe) 

data_event.groupby(data_event.columns.tolist(),as_index=False).size()

row_start, row_end = data_event.index[10], data_event.index[35] # non-unique label
row_start, row_end
# data_event.loc[row_start:row_end, 'actual':'previous']    # comment because error

#### Speeding up scalar selection

data_event.index

cn = 'Trade Balance (YoY)  (Dec)'
data_event.loc[cn, 'actual']
# data_event.loc[cn, ['actual', 'forecast', 'previous']]

data_event.at[cn, 'actual']


row_num = data_event.index.get_loc(cn)
col_num = data_event.columns.get_loc('forecast')
row_num, col_num

# %timeit data_event.iloc[row_num, col_num]

# %timeit data_event.iat[row_num, col_num]

# %timeit data_event.iloc[5, col_num]

# %timeit data_event.iloc[5, col_num]

act_val = data_event['previous']
# act_val.index

act_val.iat[50]

act_val.at['Retail Sales (MoM)  (Dec)']

#### Slicing rows lazily

data = pd.read_csv('../cookbook_data/owndata/economic_calendar_2.csv', index_col=1)
data.drop('id', axis=1, inplace=True)

time_format = '%d/%m/%Y'
data.index = pd.to_datetime(data.index, format=time_format)

events = data['event']
# events

events[4:20:4]

events.index[313]

# text - datetimeIndex issue
start = 'Mesa Community College'
stop = 'Spokane Community College'
# events[200:300:20]

first_ten_instnm = data.index[:10]
# len(first_ten_instnm), type(first_ten_instnm)

# data.loc[first_ten_instnm, ['time', 'event']]
data.loc[first_ten_instnm, ['event']]

### Slicing Lexicographically
- select data based on the lexicographic order of the values in the index
- select all rows with an index lexicographically using slice notation
- This works only if the index is sorted

pd.options.display.max_rows = 6

data = pd.read_csv('../cookbook_data/owndata/economic_calendar_2.csv', index_col=1)
data.drop('id', axis=1, inplace=True)

time_format = '%d/%m/%Y'
data.index = pd.to_datetime(data.index, format=time_format)

data_event = data.copy().set_index('event')

data_event = data_event.sort_index()

data_event.loc['AD':'AZ']

data_event = data_event.sort_index(ascending=False)
data_event.index.is_monotonic_decreasing
# college.head()

data_event.loc['C':'A']
