==> laboratory/legacy/samples/05_calendar.ipynb <==
# Boolean Indexing

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

%matplotlib inline
pd.options.display.max_columns = 10


#### Old method

# data = pd.read_csv('../cookbook_data/owndata/economic_calendar_2.csv', index_col=1)
# data.drop('id', axis=1, inplace=True)

# time_format = '%d/%m/%Y'
# data.index = pd.to_datetime(data.index, format=time_format)

#### New method

data = pd.read_csv('../cookbook_data/owndata/economic_calendar_2.csv')
# data.drop('id', axis=1, inplace=True)
data.head()

sentiments = pd.read_csv('../cookbook_data/owndata/sentiments_df.csv')
sentiments.head()

sentiments.index = sentiments['Symbol'] + '_' + sentiments['Action']
sentiments.drop(['Symbol', 'Action'], axis=1, inplace=True)
sentiments = sentiments.sort_index()
sentiments.head()

sentiments.sort_values(by=['Percentage'], ascending=False).head(6)

# sentiments.loc[['EURNZD_Long', 'NZDCHF_Short']]
sentiments.sort_values(by=['Percentage'], ascending=False).tail(6)
