==> laboratory/legacy/samples/06_shit.ipynb <==
# Index Alignment

- Index object
- Cartesian products
- Exploding object
- Filling values with unequal indexes
- Appending columns from different DataFrames
- Highlighting the maximum value from each column
- Replicating idxmax with method chaining
- Finding the most common maximum
