==> laboratory/legacy/samples/07_calendar.ipynb <==
# Grouping for Aggregation, Filtration and Transformation

import pandas as pd
import numpy as np
import pandas_ta as ta

def trend_index():
    pass


def volatility_index():
    pass


def volume_index():
    pass


def overlap_index():
    pass


def candles_pattern():
    pass


# Jan calendars
cal_1 = pd.read_csv('../cookbook_data/owndata/economic_calendar_1.csv')
cal_1.tail()

# Prices clean
price = pd.read_csv('../cookbook_data/owndata/prices_cleandf.csv')
# price.dtypes
price.tail()
# price.shape

# Sentiment combine
sent = pd.read_csv('../cookbook_data/owndata/sentiments_df.csv')
# sent.dtypes
# sent.tail()
# sent.shape

sent.sort_values(by='Percentage', ascending=False).head()

# Pound sterling xy based on GU
bxy = pd.read_csv('../cookbook_data/owndata/PHLX_British_Pound_iv.csv')
bxy.tail()

# Dollar index
dxy = pd.read_csv('../cookbook_data/owndata/US_Dollar_Index_iv.csv')
dxy.tail()

# GU fx
gu = pd.read_csv('../cookbook_data/owndata/GBPUSD_yh.csv')
gu.tail()
