==> laboratory/legacy/samples/08_sentiment.ipynb <==
import pandas as pd
import numpy as np
pd.set_option('max_columns', 10, 'max_rows', 30, 'display.precision', 4)

prices_name, sentiments_name = 'price2202_df', 'sent2202_df'

prices = pd.read_csv(f'../cookbook_data/owndata/{prices_name}.csv', index_col=0)
sentiments = pd.read_csv(f'../cookbook_data/owndata/{sentiments_name}.csv', index_col=0)

# sentiments.head()

sentiments.stack()

tidy_sentiments = sentiments.stack().reset_index()
# tidy_sentiments

tidy_sentiments.columns = ['Symbol', 'Task', 'Ret']
# tidy_sentiments

sentiments.stack().rename_axis(['Symbol', 'Task']).reset_index(name='Ret')  # chain same with above

# sentiments.melt(value_vars=['Action', 'Percentage', 'Volume', 'Positions']) # same sentiments.melt()

# sentiments.head()

# sentiments.stack()
sentiments.set_index('Action').stack()

temp = sentiments.copy()
temp.reset_index(inplace=True)
temp

# temp.melt(value_vars=['Symbol', 'Action', 'Percentage', 'Volume', 'Positions']) # same temp.melt()

temp.melt(id_vars=['Symbol'],
          value_vars=['Action', 'Percentage', 'Volume', 'Positions'])

temp.melt(id_vars=['Symbol'],
                  value_vars=['Action', 'Percentage', 'Volume', 'Positions'],
                  var_name='Cols',
                  value_name='Vals')

# temp.melt()
# temp.melt(id_vars='Symbol')
# temp.melt(id_vars='Symbol', var_name='Cols', value_name='Vals')

# sentiments_tmp = pd.read_csv(f'../cookbook_data/owndata/{sentiments_name}.csv')
# sentiments_tmp

##### sample data need "Long-Positions", "Long_Percentage", "Long_Volume"

usecol_func = lambda x: 'Positions' in x or x == 'Symbol'
sentiments_tmp = pd.read_csv(f'../cookbook_data/owndata/{sentiments_name}.csv', index_col='Symbol', usecols=usecol_func)
sentiments_stacked = sentiments_tmp.stack()
# sentiments_stacked.head(18)

sentiments_stacked.unstack().head()

# tidy_sentiments
