==> laboratory/legacy/samples/10_shit_1.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
plt.rcParams["figure.figsize"] = (20,10)
pd.set_option('max_columns', 15, 'max_rows', 30, 'display.precision', 4)

result = pd.read_csv('../cookbook_data/owndata/prices_cleandf2802.csv', index_col=0)
result_cmp = result.copy()

##### Notice: GBPJPY, GBPCAD, EURJPY, USDJPY, AUDJPY, NZDJPY

quote = r'GBP'
filter_quote = result_cmp[result_cmp.index.str.contains(quote + r'(?!$)')]
filter_quote = filter_quote.apply(pd.to_numeric)

# filter_quote

# filter_quote = result.copy()

# price_filter = filter_quote.loc[:, ['Short_Price', 'Long_Price', 'Current_Price']]
# num_chart = len(price_filter)
# # fig, axs = plt.subplots(num_chart)
# for i in range(num_chart):
#     plt.figure(figsize=(10, 6), dpi=120)
#     plt.axhline(y=price_filter.iloc[i].Short_Price, color='r', linestyle='-', label="Short_Price")
#     plt.axhline(y=price_filter.iloc[i].Long_Price, color='g', linestyle='--', label="Long_Price")
#     plt.axhline(y=price_filter.iloc[i].Current_Price, color='b', linestyle=':', label="Current_Price")
#     plt.legend(loc=4, prop={'size': 16}, title=price_filter.iloc[i].name)
#     plt.show()

# price_filter.iloc[0].name

# price_filter.iloc[0].plot()
# type(price_filter.iloc[0].Short_Price)

# plt.axhline(y=price_filter.iloc[0].Short_Price, color='r', linestyle='-', label="Short_Price")
# plt.axhline(y=price_filter.iloc[0].Long_Price, color='g', linestyle='--', label="Long_Price")
# plt.axhline(y=price_filter.iloc[0].Current_Price, color='b', linestyle=':', label="Current_Price")
# plt.legend(loc=2, prop={'size': 16}, title=price_filter.iloc[0].name)
# plt.show()

# filter_quote = result.copy()

sr = filter_quote.loc[:, ('Short_Distance', 'Long_Distance', 'Relative_SL', 'RelativePrice_SL')].mean()
sr['Combine_Rel'] = sr['Short_Distance'] / sr['Long_Distance']
sr

# sr['Combine_Rel']

filter_quote['Bed_av'] = filter_quote.Short_Distance.div(sr['Short_Distance'])
filter_quote['Bud_av'] = filter_quote.Long_Distance.div(sr['Long_Distance'])
filter_quote['Rel_avg'] = filter_quote.Relative_SL.div(sr['Relative_SL'])

filter_quote['Rel_com_avg'] = filter_quote.Relative_SL.div(sr['Combine_Rel'])

filter_quote['Rel_p_avg'] = filter_quote.RelativePrice_SL.div(sr['RelativePrice_SL'])

filter_quote = filter_quote.head(10)

filter_quote.sort_values(by='Bed_av', ascending=False, key=abs)

filter_quote.sort_values(by='Bud_av', ascending=False, key=abs)

filter_quote.sort_values(by='Rel_avg', ascending=False, key=abs)

filter_quote.sort_values(by='Rel_com_avg', ascending=False, key=abs)

filter_quote.sort_values(by='Rel_p_avg', ascending=False, key=abs)

# filter_quote.loc['GBPNZD']

sentiments_name = 'sent2802_df'

sentiments = pd.read_csv(f'../cookbook_data/owndata/{sentiments_name}.csv', index_col=0)
sent_cmp = sentiments.copy()

sent_cmp.head(10)

# sent_quote = sent_cmp[sent_cmp.index.str.contains(quote + r'(?!$)')]
# sent_quote
